requirements.txt
langchain
openai
chromadb
tiktoken
jq
python-dotenv
streamlit
langchain-community
langchain-openai
pinecone-client
langchain-pinecone
